% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_black}
\alias{theme_black}
\title{Minimalist black theme for ggplot2}
\usage{
theme_black(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Optional. Numeric. Base font size. Defaults to 11.}

\item{base_family}{Optional. Numeric. Base font family. Defaults to ?.}

\item{base_line_size}{Optional. Numeric. Base line element size.
Defaults to base_size/22.}

\item{base_rect_size}{Optional. Numeric. Base rectangle element size.
Defaults to base_size/22.}
}
\description{
A ggplot2 theme which removes most background
elements and makes all text/lines black.
}
\note{
Recommend exporting as PNG or TIFF to preserve
background transparency.
}
