% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_risk}
\alias{theme_risk}
\title{Minimalist risk table theme for ggplot2}
\usage{
theme_risk(
  base_size = 12,
  base_family = "",
  base_color = "black",
  base_line_size = base_size/14,
  base_rect_size = base_size/14
)
}
\arguments{
\item{base_size}{Optional. Numeric. Base font size. Defaults to 10.}

\item{base_family}{Optional. Numeric. Base font family. Defaults to ?.}

\item{base_color}{Optional. Character. Base color for lines and text. Defaults to black.}

\item{base_line_size}{Optional. Numeric. Base line element size.
Defaults to base_size/22.}

\item{base_rect_size}{Optional. Numeric. Base rectangle element size.
Defaults to base_size/22.}
}
\description{
A ggplot2 theme which removes most background
elements.
}
\note{
Recommend exporting as PNG or TIFF to preserve
background transparency.
}
